Allow spaces at the right cell edge while editing
Previously only one space was allowed at the right edge of a cell, but
it was also always allowed if on the line being edited. This had
undesirable side-effects such as an extra space being included at the
right edge of a cell. Now an unlimited number of spaces are allowed at
the right edge of a cell, but only while the insertion point is there
inserting them. This also makes the code simpler.